Changed data type to float to handle division

+++ server.py
@@ -100,7 +100,7 @@ def handle_source(json_data):
     data = str(json_data['data'])
     data = data[1:-1]
     print('Data before transform to np', data)
-    x = np.fromstring(data, dtype=np.int16, sep=',')
+    x = np.fromstring(data, dtype=np.float16, sep=',')
     print('data after to numpy', x)
     x = x.reshape(len(x), 96, 64, 1)
     print('Successfully reshape audio features', x.shape)
